+++ KirbyLib/CinemoKSA.cs
@@ -170,7 +170,7 @@ namespace KirbyLib
                 strings.Add(writer.BaseStream.Position, cObj.Type);
                 writer.Write(-1);
                 writer.Write(0);
-                writer.Write(writer.BaseStream.Position + 4);
+                writer.Write((uint)writer.BaseStream.Position + 4);
 
                 long varListStart = writer.BaseStream.Position;
                 writer.Write(cObj.Variables.Count);
